Block server_test until metrics server is up
In some cases it is possible that client is trying to access the server
that might not yet be up. In this case, http client returns the
'connection refused' error.
So, in case of this error, we should try to wait a bit for the metrics
server to be up.